requirements.txt
# Web UI
streamlit>=1.28.0

# TTS - ElevenLabs
elevenlabs>=1.0.0

# STT - OpenAI Whisper
openai>=1.0.0

# LLM
openai>=1.0.0

# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
